Rename nbtobject package to nbt
