+++ src/main/java/jp/mincra/mincramagics/nbt/MaterialFilter.java
@@ -1,4 +1,4 @@
-package jp.mincra.mincramagics.nbtobject;
+package jp.mincra.mincramagics.nbt;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
