+++ src/main/java/jp/mincra/mincramagics/nbt/NBTConverter.java
@@ -1,4 +1,4 @@
-package jp.mincra.mincramagics.nbtobject;
+package jp.mincra.mincramagics.nbt;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
