+++ src/main/java/jp/mincra/mincramagics/oraxen/mechanic/artifact/ArtifactMechanic.java
@@ -3,8 +3,8 @@ package jp.mincra.mincramagics.oraxen.mechanic.artifact;
 import io.th0rgal.oraxen.mechanics.Mechanic;
 import io.th0rgal.oraxen.mechanics.MechanicFactory;
 import jp.mincra.mincramagics.MincraMagics;
-import jp.mincra.mincramagics.nbtobject.Material;
-import jp.mincra.mincramagics.nbtobject.ArtifactNBT;
+import jp.mincra.mincramagics.nbt.Material;
+import jp.mincra.mincramagics.nbt.ArtifactNBT;
 import jp.mincra.mincramagics.MaterialSlot;
 import org.bukkit.configuration.ConfigurationSection;
